+++ packages/abap-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/abap-deploy-config-writer
 
+## 0.3.15
+
+### Patch Changes
+
+- Updated dependencies [c160401]
+    - @sap-ux/system-access@0.7.10
+    - @sap-ux/project-access@1.36.2
+
 ## 0.3.14
 
 ### Patch Changes